requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.28.1
anyio==4.4.0
astrapy==1.2.1
attrs==23.2.0
azure-core==1.30.2
azure-identity==1.17.0
backoff==2.2.1
beautifulsoup4==4.12.3
bitarray==2.9.2
blinker==1.8.2
boto3==1.34.129
botocore==1.34.129
bson==0.5.10
cachetools==5.3.3
cassandra-driver==3.29.1
cassio==0.1.8
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colbert-ai==0.2.19
cryptography==42.0.8
dataclasses-json==0.6.7
datasets==2.19.2
deepdiff==7.0.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
docstring_parser==0.16
emoji==2.12.1
filelock==3.15.1
filetype==1.2.0
Flask==3.0.3
frozenlist==1.4.1
fsspec==2024.3.1
geomet==0.2.1.post1
git-python==1.0.3
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.56.0
google-cloud-bigquery==3.24.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-generativeai==0.5.4
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.4
hyperframe==6.0.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.4.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
langchain==0.2.3
langchain-astradb==0.3.3
langchain-community==0.2.4
langchain-core==0.2.5
langchain-google-genai==1.0.6
langchain-google-vertexai==1.0.5
langchain-nvidia-ai-endpoints==0.1.1
langchain-openai==0.1.8
langchain-text-splitters==0.2.1
langdetect==1.0.9
langsmith==0.1.80
llama-index==0.10.43
llama-index-agent-openai==0.2.7
llama-index-cli==0.1.12
llama-index-core==0.10.43
llama-index-embeddings-azure-openai==0.1.7
llama-index-embeddings-bedrock==0.1.4
llama-index-embeddings-gemini==0.1.7
llama-index-embeddings-langchain==0.1.2
llama-index-embeddings-openai==0.1.10
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-anthropic==0.1.13
llama-index-llms-azure-openai==0.1.6
llama-index-llms-bedrock==0.1.7
llama-index-llms-gemini==0.1.10
llama-index-llms-openai==0.1.22
llama-index-llms-vertex==0.1.5
llama-index-multi-modal-llms-gemini==0.1.7
llama-index-multi-modal-llms-openai==0.1.6
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.25
llama-index-readers-llama-parse==0.1.4
llama-index-tools-cassandra==0.1.1
llama-index-vector-stores-astra-db==0.1.7
llama-index-vector-stores-cassandra==0.1.3
llama-parse==0.4.1
llamaindex-py-client==0.1.19
lxml==5.2.2
MarkupSafe==2.1.5
marshmallow==3.21.3
mpmath==1.3.0
msal==1.28.1
msal-extensions==1.1.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
nltk==3.8.1
numpy==1.26.4
openai==1.34.0
ordered-set==4.1.0
orjson==3.10.5
packaging==23.2
pandas==2.2.2
pillow==10.3.0
portalocker==2.8.2
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==14.0.1
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
PyJWT==2.8.0
pyparsing==3.1.2
pypdf==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
ragstack-ai==1.0.8
ragstack-ai-colbert==1.0.5
ragstack-ai-langchain==1.1.0
ragstack-ai-llamaindex==1.0.6
rapidfuzz==3.9.3
regex==2024.5.15
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
s3transfer==0.10.1
safetensors==0.4.3
scipy==1.13.1
setuptools==70.0.0
shapely==2.0.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
striprtf==0.0.26
sympy==1.12.1
tabulate==0.9.0
tenacity==8.4.1
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
torch==2.2.1
tqdm==4.66.4
transformers==4.41.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
unstructured==0.14.2
unstructured-client==0.23.7
uritemplate==4.1.1
urllib3==2.2.2
uuid6==2024.1.12
Werkzeug==3.0.3
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
